# CDAP-3902: Remove unnecessary remote calls when getting a dataset instance