--- a/cdap-data-fabric/src/main/java/co/cask/cdap/data/runtime/DataSetsModules.java
+++ b/cdap-data-fabric/src/main/java/co/cask/cdap/data/runtime/DataSetsModules.java
@@ -18,6 +18,8 @@ package co.cask.cdap.data.runtime;
 
 import co.cask.cdap.api.dataset.module.DatasetDefinitionRegistry;
 import co.cask.cdap.common.runtime.RuntimeModule;
+import co.cask.cdap.data2.datafabric.dataset.DatasetProvider;
+import co.cask.cdap.data2.datafabric.dataset.DefaultDatasetProvider;
 import co.cask.cdap.data2.datafabric.dataset.RemoteDatasetFramework;
 import co.cask.cdap.data2.datafabric.dataset.type.DatasetTypeClassLoaderFactory;
 import co.cask.cdap.data2.datafabric.dataset.type.DistributedDatasetTypeClassLoaderFactory;
@@ -46,6 +48,9 @@ public class DataSetsModules extends RuntimeModule {
                   .build(DatasetDefinitionRegistryFactory.class));
         bind(DatasetFramework.class).to(InMemoryDatasetFramework.class).in(Scopes.SINGLETON);
         expose(DatasetFramework.class);
+
+        bind(DatasetProvider.class).to(DefaultDatasetProvider.class);
+        expose(DatasetProvider.class);
       }
     };
   }
@@ -62,6 +67,9 @@ public class DataSetsModules extends RuntimeModule {
         expose(DatasetTypeClassLoaderFactory.class);
         bind(DatasetFramework.class).to(RemoteDatasetFramework.class);
         expose(DatasetFramework.class);
+
+        bind(DatasetProvider.class).to(DefaultDatasetProvider.class);
+        expose(DatasetProvider.class);
       }
     };
 
@@ -79,6 +87,9 @@ public class DataSetsModules extends RuntimeModule {
         expose(DatasetTypeClassLoaderFactory.class);
         bind(DatasetFramework.class).to(RemoteDatasetFramework.class);
         expose(DatasetFramework.class);
+
+        bind(DatasetProvider.class).to(DefaultDatasetProvider.class);
+        expose(DatasetProvider.class);
       }
     };
   }