--- a/cdap-data-fabric/src/main/java/co/cask/cdap/data2/datafabric/dataset/RemoteDatasetFramework.java
+++ b/cdap-data-fabric/src/main/java/co/cask/cdap/data2/datafabric/dataset/RemoteDatasetFramework.java
@@ -21,24 +21,19 @@ import co.cask.cdap.api.dataset.DatasetAdmin;
 import co.cask.cdap.api.dataset.DatasetContext;
 import co.cask.cdap.api.dataset.DatasetProperties;
 import co.cask.cdap.api.dataset.DatasetSpecification;
-import co.cask.cdap.api.dataset.module.DatasetDefinitionRegistry;
 import co.cask.cdap.api.dataset.module.DatasetModule;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.common.lang.ClassLoaders;
 import co.cask.cdap.common.utils.ApplicationBundler;
 import co.cask.cdap.data2.datafabric.dataset.type.DatasetTypeClassLoaderFactory;
 import co.cask.cdap.data2.dataset2.DatasetDefinitionRegistryFactory;
 import co.cask.cdap.data2.dataset2.DatasetFramework;
 import co.cask.cdap.data2.dataset2.DatasetManagementException;
+import co.cask.cdap.data2.dataset2.InstanceConflictException;
 import co.cask.cdap.data2.dataset2.SingleTypeModule;
-import co.cask.cdap.data2.dataset2.module.lib.DatasetModules;
 import co.cask.cdap.proto.DatasetMeta;
-import co.cask.cdap.proto.DatasetModuleMeta;
 import co.cask.cdap.proto.DatasetSpecificationSummary;
 import co.cask.cdap.proto.DatasetTypeMeta;
 import co.cask.cdap.proto.Id;
-import com.google.common.base.Objects;
-import com.google.common.base.Throwables;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
@@ -55,7 +50,6 @@ import org.slf4j.LoggerFactory;
 import java.io.File;
 import java.io.IOException;
 import java.util.Collection;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.jar.JarEntry;
@@ -71,8 +65,7 @@ public class RemoteDatasetFramework implements DatasetFramework {
   private static final Logger LOG = LoggerFactory.getLogger(RemoteDatasetFramework.class);
 
   private final LoadingCache<Id.Namespace, DatasetServiceClient> clientCache;
-  private final DatasetDefinitionRegistryFactory registryFactory;
-  private final DatasetTypeClassLoaderFactory typeLoader;
+  private final AbstractDatasetProvider instances;
 
   @Inject
   public RemoteDatasetFramework(final DiscoveryServiceClient discoveryClient,
@@ -84,8 +77,23 @@ public class RemoteDatasetFramework implements DatasetFramework {
         return new DatasetServiceClient(discoveryClient, namespace);
       }
     });
-    this.registryFactory = registryFactory;
-    this.typeLoader = typeLoader;
+    this.instances = new AbstractDatasetProvider(registryFactory, typeLoader) {
+      @Override
+      public DatasetMeta getMeta(Id.DatasetInstance instance) throws Exception {
+        return RemoteDatasetFramework.this.clientCache.getUnchecked(instance.getNamespace())
+          .getInstance(instance.getId());
+      }
+
+      @Override
+      public void createIfNotExists(Id.DatasetInstance instance, String type,
+                                    DatasetProperties creationProps) throws Exception {
+        try {
+          RemoteDatasetFramework.this.addInstance(type, instance, creationProps);
+        } catch (InstanceConflictException e) {
+          // ignore, since this indicates the dataset already exists
+        }
+      }
+    };
   }
 
   @Override
@@ -186,7 +194,7 @@ public class RemoteDatasetFramework implements DatasetFramework {
       return null;
     }
 
-    DatasetType type = getDatasetType(instanceInfo.getType(), classLoader);
+    DatasetType type = instances.getDatasetType(instanceInfo.getType(), classLoader);
     return (T) type.getAdmin(DatasetContext.from(datasetInstanceId.getNamespaceId()), instanceInfo.getSpec());
   }
 
@@ -202,9 +210,8 @@ public class RemoteDatasetFramework implements DatasetFramework {
       return null;
     }
 
-    DatasetType type = getDatasetType(instanceInfo.getType(), classLoader);
-    return (T) type.getDataset(DatasetContext.from(datasetInstanceId.getNamespaceId()),
-                               instanceInfo.getSpec(), arguments);
+    return instances.get(datasetInstanceId, instanceInfo.getType(), instanceInfo.getSpec(),
+                         classLoader, arguments);
   }
 
   @Override
@@ -301,51 +308,8 @@ public class RemoteDatasetFramework implements DatasetFramework {
 
   // can be used directly if DatasetTypeMeta is known, like in create dataset by dataset ops executor service
   public <T extends DatasetType> T getDatasetType(DatasetTypeMeta implementationInfo,
-                                                  ClassLoader classLoader)
-    throws DatasetManagementException {
+                                                  ClassLoader classLoader) throws DatasetManagementException {
 
-
-    if (classLoader == null) {
-      classLoader = Objects.firstNonNull(Thread.currentThread().getContextClassLoader(), getClass().getClassLoader());
-    }
-
-    DatasetDefinitionRegistry registry = registryFactory.create();
-    List<DatasetModuleMeta> modulesToLoad = implementationInfo.getModules();
-    for (DatasetModuleMeta moduleMeta : modulesToLoad) {
-      // adding dataset module jar to classloader
-      try {
-        classLoader = typeLoader.create(moduleMeta, classLoader);
-      } catch (IOException e) {
-        LOG.error("Was not able to init classloader for module {} while trying to load type {}",
-                  moduleMeta, implementationInfo, e);
-        throw Throwables.propagate(e);
-      }
-
-      Class<?> moduleClass;
-
-      // try program class loader then cdap class loader
-      try {
-        moduleClass = ClassLoaders.loadClass(moduleMeta.getClassName(), classLoader, this);
-      } catch (ClassNotFoundException e) {
-        try {
-          moduleClass = ClassLoaders.loadClass(moduleMeta.getClassName(), null, this);
-        } catch (ClassNotFoundException e2) {
-          LOG.error("Was not able to load dataset module class {} while trying to load type {}",
-                    moduleMeta.getClassName(), implementationInfo, e);
-          throw Throwables.propagate(e);
-        }
-      }
-
-      try {
-        DatasetModule module = DatasetModules.getDatasetModule(moduleClass);
-        module.register(registry);
-      } catch (Exception e) {
-        LOG.error("Was not able to load dataset module class {} while trying to load type {}",
-                  moduleMeta.getClassName(), implementationInfo, e);
-        throw Throwables.propagate(e);
-      }
-    }
-
-    return (T) new DatasetType(registry.get(implementationInfo.getName()), classLoader);
+    return instances.getDatasetType(implementationInfo, classLoader);
   }
 }