https://issues.cask.co/browse/CDAP-3902
http://builds.cask.co/browse/CDAP-RBT518

- [x] Integration tests (NamespaceTest,DatasetTest,StreamTest)
- [x] Manually test purchase example

Merged to develop (3.3): https://github.com/caskdata/cdap/pull/4325
Merged to release/3.1 (3.2.1): https://github.com/caskdata/cdap/pull/4340

---

# Performance

## Config

```
    <property>
      <name>dataset.service.exec.threads</name>
      <value>300</value>
    </property>
    <property>
      <name>dataset.service.worker.threads</name>
      <value>30</value>
    </property>
    <property>
      <name>app.exec.threads</name>
      <value>20</value>
    </property>
```

## Test results

N threads, each making 1 get dataset request with 1 owner

*Without fix*
N=100: ~11000 ms
N=300: ~139000 ms
N=500: ~230000 ms
N=1000: ~441000 ms

*With fix*
N=100: ~8000 ms
N=300: ~123000 ms
N=500: ~195000 ms
N=1000: ~444000 ms