@poornachandra 2015-10-29
@jwang47 Can you reference the PR from which you are cherry-picking this commit? It will be easier to review then.