@jwang47 2015-10-29
This isn't cherry-picked, as there were too many merge conflicts.