@poornachandra 2015-10-31
@jwang47 One more question, otherwise changes look good.

Can you run integration tests for this change?