@poornachandra 2015-10-31
Can you also add the performance numbers in this PR?