@jwang47 2015-10-31
Integration tests (NamespaceTest,DatasetTest,StreamTest) passed. Will work on performance numbers.