@jwang47 2015-11-02
Performance numbers look OK, going to merge.